requirements.txt
streamlit==1.28.2
bitrix24-rest==2.0.3
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
plotly==5.17.0